ICU and Ultrasonic and some improvement
* software ICU driver 
*Using ICU with Ultrasonic sensor " US_Applicati_Main.c"
*Fixed and complete Timer driver 
*Fixed gpio driver  "gpioPinDirection &gpioPinWrite" function

+++ timers.h
@@ -180,7 +180,7 @@ void timer0DelayUs(uint32_t u32_delay_in_us);
  * Description:
  * @param dutyCycle
  */
-void timer0SwPWM(uint8_t u8_dutyCycle,uint32_t u8_frequency);
+void timer0SwPWM(uint8_t u8_dutyCycle,uint8_t u8_frequency);
